DataAugmentation_Projekt/requirements.txt
opencv-python
pillow
tqdm
numpy